+++ test/TokenList.test.js
@@ -1,5 +1,5 @@
-import ether from 'openzeppelin-solidity/test/helpers/ether';
-import expectThrow from 'openzeppelin-solidity/test/helpers/expectThrow';
+import { ether } from 'openzeppelin-solidity/test/helpers/ether';
+import { reverting } from 'openzeppelin-solidity/test/helpers/shouldFail';
 import { inLogs } from 'openzeppelin-solidity/test/helpers/expectEvent';
 
 const BigNumber = web3.BigNumber;
@@ -35,7 +35,7 @@ contract('Token', function ([_, user, owner, tokenWallet, tokenContract, token2]
         it('when existed', async function () {
             await this.Dinngo.addToken(id, this.Token.address, { from: owner });
             let tokenAddress1 = await this.Dinngo.tokenID_Address.call(2);
-            await expectThrow(this.Dinngo.addToken(id2, this.Token.address, { from: owner }));
+            await reverting(this.Dinngo.addToken(id2, this.Token.address, { from: owner }));
             await this.Dinngo.addToken(id2, token2, { from: owner });
             let tokenAddress2 = await this.Dinngo.tokenID_Address.call(3);
             tokenAddress2.should.eq(token2);
@@ -53,14 +53,14 @@ contract('Token', function ([_, user, owner, tokenWallet, tokenContract, token2]
         });
 
         it('when token not exist', async function () {
-            await expectThrow(this.Dinngo.updateTokenRank(this.Token.address, 2, { from: owner }));
+            await reverting(this.Dinngo.updateTokenRank(this.Token.address, 2, { from: owner }));
         });
 
         it('when assigning same rank', async function () {
             await this.Dinngo.addToken(2, this.Token.address, { from: owner });
             let rank1 = await this.Dinngo.tokenRanks.call(this.Token.address);
             rank1.should.be.bignumber.eq(1);
-            await expectThrow(this.Dinngo.updateTokenRank(this.Token.address, 1, { from: owner }));
+            await reverting(this.Dinngo.updateTokenRank(this.Token.address, 1, { from: owner }));
         });
     });
 });