+++ test/UserList.test.js
@@ -1,5 +1,5 @@
-import ether from 'openzeppelin-solidity/test/helpers/ether';
-import expectThrow from 'openzeppelin-solidity/test/helpers/expectThrow';
+import { ether } from 'openzeppelin-solidity/test/helpers/ether';
+import { reverting } from 'openzeppelin-solidity/test/helpers/shouldFail';
 import { inLogs } from 'openzeppelin-solidity/test/helpers/expectEvent';
 
 const BigNumber = web3.BigNumber;
@@ -34,7 +34,7 @@ contract('User', function ([_, user, owner, tokenWallet, tokenContract, user2, s
         it('when existed', async function () {
             await this.Dinngo.addUser(id, user, { from: owner });
             let userAddress1 = await this.Dinngo.userID_Address.call(id);
-            await expectThrow(this.Dinngo.addUser(id2, user, { from: owner }));
+            await reverting(this.Dinngo.addUser(id2, user, { from: owner }));
             let userAddress2 = await this.Dinngo.userID_Address.call(id2);
             userAddress1.should.eq(user);
             userAddress2.should.eq(ZERO_ADDRESS);
@@ -55,14 +55,14 @@ contract('User', function ([_, user, owner, tokenWallet, tokenContract, user2, s
         });
 
         it('when user not exist', async function () {
-            await expectThrow(this.Dinngo.updateUserRank(user, 2, { from: owner }));
+            await reverting(this.Dinngo.updateUserRank(user, 2, { from: owner }));
         });
 
         it('when assigning same rank', async function () {
             await this.Dinngo.addUser(id, user, { from: owner });
             let rank1 = await this.Dinngo.userRanks.call(user);
             rank1.should.be.bignumber.eq(1);
-            await expectThrow(this.Dinngo.updateUserRank(user, 1, { from: owner }));
+            await reverting(this.Dinngo.updateUserRank(user, 1, { from: owner }));
         });
     });
 
@@ -78,11 +78,11 @@ contract('User', function ([_, user, owner, tokenWallet, tokenContract, user2, s
 
         it('when call by non owner', async function() {
             await this.Dinngo.setUser(id, user, 1);
-            await expectThrow(this.Dinngo.removeUser(user, { from: someone }));
+            await reverting(this.Dinngo.removeUser(user, { from: someone }));
         });
 
         it('when removing non-existed user', async function() {
-            await expectThrow(this.Dinngo.removeUser(user, { from: owner }));
+            await reverting(this.Dinngo.removeUser(user, { from: owner }));
         });
     });
 });